+++ src/main/java/warmup/one/WarmupOneFourthSet.java
@@ -9,7 +9,7 @@ public class WarmupOneFourthSet {
      *
      * @param str
      * @param n
-     * @return
+     * @return String
      */
     public String missingChar(String str, int n) {
         if (n == 0) {
@@ -23,7 +23,7 @@ public class WarmupOneFourthSet {
      * Given a string, return a new string where the first and last chars have been exchanged.
      *
      * @param str
-     * @return
+     * @return String
      */
     public String frontBack(String str) {
         if (str.length() == 0 || str.length() == 1) {
@@ -35,5 +35,24 @@ public class WarmupOneFourthSet {
         String otherCharacters = str.substring(1, str.length()-1);
 
         return lastCharacter + otherCharacters + firstCharacter;
+    }
+
+    /**
+     * Given a string, we'll say that the front is the first 3 chars of the string. If the string length
+     * is less than 3, the front is whatever is there. Return a new string which is 3 copies of the front.
+     *
+     * @param str
+     * @return String
+     */
+    public String front3(String str) {
+        String resultString = "";
+
+        if (str.length() < 3) {
+            return str + str + str;
+        }
+
+        resultString = str.substring(0, 3);
+
+        return resultString + resultString + resultString;
     }
 }